Have single Run function for p2p code
The run function connects to peers, joins topics, and listens on all
topics. Previously, the caller had to listen on the individual topics.
This became problematic with the unified run function as they sometimes
subscribed before joining the topic was complete.

All received messages are sent to a single channel. This interface was
easiest to implement for now, we can later change it.