Make eon public key deterministic
The mocknode now uses the same eon public key for all sent cipher
batches. It derives it from a seed value stored in the config and prints
it at startup. For testing, this key should be inserted into the
decryptor db so that they can verify the decryption keys they receive.