Align config file headers
For convenience, we put addresses and public keys that are derived from
private keys as comments in the config files. This allows us to extract
them via a regex during testing. Before this commit, the keyper used a
format different from the one used by keypers or the collator, which
lead to ambiguities.