Add retryGetAddrs function
This allows us to shorten the eventHandler's methods a bit.
I'll only this in the decryptor's eventhandling, since that will become the
shared eventhandling code eventually.

One other problem we have here, is that we try to get those addresses while
we're running inside a database transaction. We need to fix that later.