Add HTTP GET endpoint for signed EonPublicKey messages
Nodes that want to encrypt transactions with the correct EonPublicKey
need a way to retrieve the signed EonPublicKey gossip messages
from the Keypers. Since nodes are not expected to be online constantly,
they should not listen to broadcast messages.

The Collator persists the EonPublicKey messages and makes them
accessible for the nodes in byte-encoded protobuf form.
Nodes should decode and verify the signature of all messages without
trusting the Collator.
All information (including the correct expected signer-Address) should
be queried from the blockchain and checked for equality in all messages.