Add validation and handling of DecryptionKey in Collator
The collator now receives DecryptionKey messages from Keypers.
Upon validation of each message, a matching EonPublicKey with enough
confirmations is retrieved from the Database (inserted from valid
EonPublicKey keyper messages). The Epoch-Key is verified against the
Public-key.

The handler of the DecryptionKey will insert the (validated)
epoch-key into the Collator's database for later retrieval.

The triggering of the decryption process (on the rollup) is not
implemented in this commit.