Let keypers vote on the success of the DKG process
The keypers now vote on the success of the DKG process. If a majority fails, the
shuttermint app restarts the DKG process. This replaces the EonStart vote that
they previously sent if the DKG process failed.

Instead of voting on the config index or activiation block, we now also
reference the eon for which we started the DKG process.

See https://github.com/shutter-network/rolling-shutter/issues/250