Split the definition of the eon public key message
This adds the EonPublicKeyCandidate protocol buffer type, which contains the
part of the message which the keypers sign.

One important change is that the keypers do not sign the keyper index
anymore. As a result all of the keypers sign the same data, and we will be able
to take advantage of that later (e.g. we now have a canonical database key for a
candidate).