Add EpochTicker broker
`EpochTicker` uses the `Broker` subscription class
in order to distribute a ticker signal to multiple
subscribers.
This will get used to feed events to pending batches
that trigger the creation of the DecryptionTrigger
as soon as a new Epoch should be initiated, based
on a constant time interval.

Unlike the `time.Ticker` there is still one problem
with slow receivers:

Since consumers subscribe to the ticker
with a non-blocking send,it is possible to miss a tick
when the receiving loop is busy.
Unlike in the actual ticker, the consumer now has to wait until the
next tick is reached.