Add Batch run method
The batches `Run` method is the main go-routine that will run
during the lifetime of that `Batch`.
It receives the incoming events from the public channels
and dispatches it to the current `State` object for that batch.
It check's and handles the state-transitions and emits the
StateChangeResult of the state-transitions via the batches `Broker`.