Remove unused threadsafety feature of sequencer.Cached
We decided that the mutexes used for the per-address locking
mechanism was too unpredictable in behavior and the class
itself only ever will be used from one go-routine non-concurrently.
Thus the threadsafety of the Cached class was removed completely.

Beware that from now on this class should never be used
concurrently!