feat(mocksequencer): add support for JSON RPC error code injection
The methods of the RPCService implementations could return
errors, which were converted to the approriate JSON-RPC response
value in the rpc library, e.g.:

```json
{
    "id": 1337
    "jsonrpc": "2.0",
    "error": {
        "code": -32003,
        "message": "Transaction rejected"
    }
}
```

However, there was no possibility to specify the error-code,
which follows general JSON-RPC 2.0 specs as well as EIP-1474 and
EIP-1193.
All errors were automatically encoded with the default error-code
`-32000`.

Now it is possible to inject specific error-codes anywhere in the
"error-stack" (pkg/errors wrap calls), that can later be retrieved
and converted to an error that complies with the `go-ethereum/rpc.Error`
interface.

If an RPC error was injected not on the topmost level of the
"error stack", a call to `ExtractRPCError` has to be made
before returning the error in the RPCService methods.