feat(collator): add batch-production routine to collator
closes #335

A new routine is added which closes batches at
a constant interval by calling `batcher.CloseBatch()`
and afterwards trying to initiate the batchers
chain-state by spamming `batcher.EnsureChainState()`.
We do this early so that eventually queued up
transactions will be validated before the batch is
being close again.
This is favourable to do as early as possible during
the batch life-cycle, since we query account state
from the sequencer for the validation.

All calls to the batcher are retried indefinitely until
successful.
Canceling the passed in `ctx` will cancel all retries
and batch-closing attempts.

The routine makes some assumptions about timings:

- It assumes a constant time between closing the
  batch and trying to initiate the chain-state.
  This is the time it takes until the batch is committed to,
  decrypted and processed by the sequencer.

- The retry-interval is a fraction of the
  overall `interval` argument and is capped at a
  constant minimal time.

What is not yet considered in this commit:

- In the future the assumed timing parameters should be made
  configurable and dynamically scalable.

- In the future the failing error conditions should be
  extended, since there might be some errors in the
  call stacks that are not recoverable and should exit
  the collator process.