feat(mocksequencer): add minimal GetBlockByNumber endpoint
The collator queries the sequencer for the current block and
uses the GasLimit and the BaseFee for batch creation.

Since the eth-client library validates required JSON fields of the
block, the structure of the returned block in the response
has to comply to this.
Therefore all required fields are returned with their null-values,
even if the collator does not process them.