fix(shuttermint): additional super-majority votes triggered DKG
Fixes #348

A check is added that disregards super-majority votes for previously
processed DKG outcome votes.

This fixes the previous behavior:

During a finalized DKG process, the keypers vote whether the DKG
was successful or not. A majority is reached as soon as the
threshold of keypers agree on one outcome.

When the DKG vote of the threshold-completing keyper is processed
with outcome `success=false`, the `maybeStartEon()` method evaluates
to true and starts a new DKG process as expected.
This will increase the internal `app.EONCounter`.

However, when additional super-majority DKG votes for the same
(now outdated) DKG are processed,
the `maybeStartEon()` method keeps evaluating to true
and will again start a new DKG process, increasing the internal
counter further.
This will keep happening for each outdated super-majority DKG that is
processed.