feat: add options in message validation
This commit changes the interface for the
registration of the topic-validators that
are passed to gossipsub.

It is now possible to add more fine-grained
control over the action taken when we
receive an invalid message from a peer.

Changes to `AddValidator`:

When providing the optional `WithGracefulIgnore`
option, invalid messages (when the handler function
returns false) will not result in lowering the
score of the peer, but simply ignoring the message.

When providing the optional `WithTraceContextPropagation`
option, the handlers will accept envelope-messages
with non-empty optional `trace` field.
This option is false per default in order to discourage
peers to send any unnecessary extra-data in production
environments.
In testing environments this needs to be enabled
when trace-context propagation is desired.

The AddValidator registration function now returns
a possible error value instead of the handler
function implementing the gossipsub Validator interface.