feat: add opentelemetry tracing and instrumentation
- Injection and Extraction of otel trace-context

In order to correlate spans/traces across process/peer
boundaries, otel needs to inject/extract remote trace-context
metadata from the gossipsub messaging layer and attach/
detach that to the passed `context.Context` instance.
This is now conducted in the `p2p.SendMessage` and
`p2p.handle` methods respectively.
The trace-context is carried via the optional `trace`
field in the message `Envelope`.

- Semi-automatic instrumentation

The spans to be collected by the tracing collectors
still have to be manually started and ended
("manual instrumentation").
We provide a convenience function `trace.StartSpan`
that eventually starts a span, injects it in the context
and already attaches the most relevant information
from callstack-introspection.
The returned span-instance still has to be ended
manually (see the function's documentation).

Usually that will be already enough to gain enough
insight in the programs spans, but additional
attributes can still be attached to the span.

If tracing is not enabled (e.g. in production-environments),
the function will only conduct noop operations - so it is
safe to call throughout the code without explicitly checking
wether tracing is enabled.
(Beware that this adds a small amount of call-overhead)

- Exposing trace/metrics collector endpoints

The program has to expose endpoints where the tracing
and metrics collectors can poll for new traces or metrics.
Additionally some global variables have to be set in order
to allow for easy instrumentation throughout the codebase.
This is handled by running the `trace.Run` service from
the very outer entrypoints, e.g. in the cmd/main functions.